refactor: Move the cancel app into its own isolated express app
This change follows the pattern we've started to use for the "new" move
app and the person escort record/youth risk assessment.

It isolates the logic around cancelling a move. This should help with
maintaining the move application as it has grown considerably.
