+++ app/move/app/cancel/steps.js
@@ -1,4 +1,4 @@
-const { Cancel } = require('../controllers')
+const { CancelController } = require('./controllers')
 
 module.exports = {
   '/': {
@@ -9,11 +9,11 @@ module.exports = {
   },
   '/reason': {
     backLink: null,
-    template: 'move/views/cancel',
+    template: 'move/app/cancel/cancel',
     pageTitle: 'moves::cancel.steps.reason.heading',
     buttonText: 'actions::confirm_cancellation',
     buttonClasses: 'govuk-button--warning',
     fields: ['cancellation_reason', 'cancellation_reason_other_comment'],
-    controller: Cancel,
+    controller: CancelController,
   },
 }
