+++ app/move/config/index.js
@@ -1,11 +1,9 @@
 const assignConfig = require('./assign.config')
-const cancelConfig = require('./cancel.config')
 const unassignConfig = require('./unassign.config')
 const updateConfig = require('./update.config')
 
 module.exports = {
   assignConfig,
-  cancelConfig,
   unassignConfig,
   updateConfig,
 }
