+++ app/move/controllers/index.js
@@ -1,5 +1,4 @@
 const assign = require('./assign')
-const Cancel = require('./cancel')
 const confirmation = require('./confirmation')
 const timeline = require('./timeline')
 const Unassign = require('./unassign')
@@ -8,7 +7,6 @@ const view = require('./view')
 
 module.exports = {
   assign,
-  Cancel,
   confirmation,
   timeline,
   Unassign,
