+++ app/move/fields/index.js
@@ -1,9 +1,6 @@
 const { cloneDeep } = require('lodash')
 
 const additionalInformation = require('./additional-information')
-const cancellationReason = require('./cancellation-reason')
-const cancellationReasonCancelledByPMUComment = require('./cancellation-reason-cancelled-by-pmu-comment')
-const cancellationReasonOtherComment = require('./cancellation-reason-other-comment')
 const assessmentAnswer = require('./common.assessment-answer')
 const courtHearingComments = require('./court-hearing-comments')
 const courtHearingCourtCase = require('./court-hearing-court-case')
@@ -47,12 +44,6 @@ const toLocationPrisonTransfer = require('./to-location-prison-transfer')
 const toLocationSecureChildrensHomeTransfer = require('./to-location-secure-childrens-home')
 const toLocationSecureTrainingCentreTransfer = require('./to-location-secure-training-centre')
 const videoRemandComments = require('./video-remand-comments')
-
-const cancelFields = {
-  cancellation_reason: cancellationReason,
-  cancellation_reason_cancelled_by_pmu_comment: cancellationReasonCancelledByPMUComment,
-  cancellation_reason_other_comment: cancellationReasonOtherComment,
-}
 
 const createFields = {
   additional_information: additionalInformation,
@@ -133,7 +124,6 @@ const assignFields = {
 
 module.exports = {
   assignFields,
-  cancelFields,
   createFields,
   updateFields,
 }
