+++ app/move/steps/index.js
@@ -1,11 +1,9 @@
 const assign = require('./assign')
-const cancel = require('./cancel')
 const unassign = require('./unassign')
 const update = require('./update')
 
 module.exports = {
   assign,
-  cancel,
   unassign,
   update,
 }
